refactor: combat.py에서 Combatant/CombatAction/CombatantType을 combatant.py로 분리
- combatant.py (112줄): CombatAction, CombatantType, Combatant
- combat.py (414줄): CombatTurn, CombatInstance + re-export (하위 호환성)
